% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{countsNorm}
\alias{countsNorm}
\title{countsNorm}
\arguments{
\item{counts}{A counts matrix.}
}
\description{
countsNorm
}
\examples{

\dontrun{countsNorm()}

}
\author{
Jason T. Serviss
}
\keyword{countsNorm}
