% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{doMerge}
\alias{doMerge}
\title{doMerge}
\arguments{
\item{counts.norm;}{Matrix with normalized expression values.}

\item{class}{A tibble with columns \emph{class} and \emph{sample} indicating
the class ID and sample ID respectivley.}
}
\description{
doMerge
}
\author{
Jason T. Serviss
}
\keyword{doMerge}
