% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{doSplit}
\alias{doSplit}
\title{doSplit}
\arguments{
\item{counts.norm;}{Matrix with normalized expression values.}

\item{class}{A tibble with columns \emph{class} and \emph{sample} indicating
the class ID and sample ID respectivley.}

\item{iter}{A numeric vector of length 1 indicating the number of random
splits to test.}

\item{subset}{A character vector of length 1 indicating the specific class to
investigate.}
}
\description{
doSplit
}
\author{
Jason T. Serviss
}
\keyword{doSplit}
