% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{foldChangePerGroup}
\alias{foldChangePerGroup}
\title{foldChangePerGroup}
\arguments{
\item{mat}{The matrix holding expression values.}

\item{class}{A tibble with columns \emph{class} and \emph{sample} indicating
the class ID and sample ID respectivley.}

\item{subset}{An optional character vector in the case that the fold change
for only a subset of groups should be calculated.}
}
\description{
foldChangePerGroup
}
\author{
Jason T. Serviss
}
\keyword{foldChangePerGroup}
