% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{groupNames}
\alias{groupNames}
\title{groupNames}
\arguments{
\item{groups.fetal}{Character vector with Martin's groupings of the cells.}
}
\description{
groupNames
}
\examples{

\dontrun{groupNames()}

}
\author{
Jason T. Serviss
}
\keyword{groupNames}
