% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{meanPerGroup}
\alias{meanPerGroup}
\title{meanPerGroup}
\arguments{
\item{mat}{The matrix holding expression values.}

\item{class}{A tibble with columns \emph{class} and \emph{sample} indicating
the class ID and sample ID respectivley.}
}
\description{
meanPerGroup
}
\author{
Jason T. Serviss
}
\keyword{meanPerGroup}
