% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{scaleFeatures}
\alias{scaleFeatures}
\title{scaleFeatures}
\arguments{
\item{mat}{The marrix to be scaled.}
}
\description{
Scales rows of a matrix between 0 and 1.
}
\examples{

mat <- matrix(1:100, ncol = 10, nrow = 10)
scaleFeatures(mat)

}
\author{
Jason T. Serviss
}
\keyword{scaleFeatures}
