% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{scaleFeaturesV}
\alias{scaleFeaturesV}
\title{scaleFeaturesV}
\arguments{
\item{v}{The marrix to be scaled.}
}
\description{
Scales a vector between 0 and 1.
}
\examples{

v <- 1:100
scaleFeaturesV(v)

}
\author{
Jason T. Serviss
}
\keyword{scaleFeaturesV}
